fix data type

+++ apps/bfd-pipeline-idr/model2/idr_claim_item_professional_ss.py
@@ -183,8 +183,8 @@ class IdrClaimItemProfessionalSs(IdrBaseModel):
     clm_line_rbndlg_crtfctn_num: Annotated[
         str, {ALIAS: ALIAS_LINE_MCS}, BeforeValidator(transform_default_string)
     ]
-    clm_line_hct_lvl_num: Annotated[int | None, {ALIAS: ALIAS_LINE_MCS}]
-    clm_line_hgb_lvl_num: Annotated[int | None, {ALIAS: ALIAS_LINE_MCS}]
+    clm_line_hct_lvl_num: Annotated[float | None, {ALIAS: ALIAS_LINE_MCS}]
+    clm_line_hgb_lvl_num: Annotated[float | None, {ALIAS: ALIAS_LINE_MCS}]
     idr_insrt_ts_line_mcs: Annotated[
         datetime,
         {ALIAS: ALIAS_LINE_MCS, **INSERT_FIELD},
